// src/app/core/models/informacoes-complementares/informacoes-complementares-request.ts
export interface InformacoesComplementaresRequest {
  dataNascimento?: string;
  nomeEmpresa?: string;
  profissao?: string;
  atividadeEconomica?: string;
  estadoCivil?: string;
  codigo?: string;
  nomePai?: string;
  nomeMae?: string;
  naturalidade?: string;
  nacionalidade?: string;
  comentario?: string;
}